PlugAndPlay E2E test cleanup (#4064)
This PR cleans up workarounds that were put in when the PlugAndPlay tests were first made.

Previously, we needed to use a preview version of IoTHub that supported PnP. Since then, they've gone GA, and we can use our normal IoTHub and normal EventHub endpoint. We can also get rid of the HTTP request we manually made and use the ServiceSDK's GetTwinAsync call that now contains the ModelID.